nros_rmw_zenoh::shim::service

Struct ZenohServiceServer

pub struct ZenohServiceServer { /* private fields */ }

Zenoh service server using queryables

Receives service requests via queryable callbacks. Note: The reply mechanism is limited due to the callback model.

Implementations

impl ZenohServiceServer

pub fn new( context: &Context, service: &ServiceInfo<'_>, liveliness: Option<LivelinessToken>, ) -> Result<Self, TransportError>

Create a new service server for the given service